Fix favorites status synchronization

+++ app/src/main/java/br/com/regmoraes/marvelcharacters/model/Comic.kt
@@ -1,8 +1,7 @@
 package br.com.regmoraes.marvelcharacters.model
 
 data class Comic(
-    val id: Int,
+    val id: Long,
     val title: String,
     val thumbnail: Thumbnail
 )
-
